pyproject.toml
[tool.poetry]
name = "presign"
version = "0.0.1"
description = ""
authors = ["pajowu <git@ca.pajowu.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
django = ">=4.1.5"
django-allauth = "^0.52.0"
pyyaml = "^6.0"
django-bootstrap5 = "^22.2"
django-compressor = "^4.3.1"
django-libsass = "^0.9"
django-scopes = "^1.2.0.post1"
django-phonenumber-field = { extras = ["phonenumbers"], version = "^7.0.2" }
django-debug-toolbar = "^3.8.1"
django-clone = "^5.3.0"
django-simple-history = "^3.2.0"
django-livereload-server = "^0.4"
django-i18nfield = "^1.9.4"
django-configurations = "^2.4"
django-hierarkey = "^1.1.0"
markdown = "^3.4.1"
nh3 = "^0.2.8"


[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.11"
djlint = "^1.19.13"
pillow = "^9.4.0"
pytest = "^7.2.1"
pytest-django = "^4.5.2"
pytest-playwright = "^0.3.0"
pytest-cov = "^4.0.0"
pytest-factoryboy = "^2.5.1"
django-eml-email-backend = "^0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
profile = "black"
src_paths = "presign"
default_section = "THIRDPARTY"
known_django = "django"
sections = "FUTURE,STDLIB,DJANGO,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"

[tool.djlint]
preserve_blank_lines = true
ignore = "T003,H006,H030,H031"
